# fix: iteratively replace "close" to avoid maximum stack error