- Closes https://github.com/alexeyraspopov/picocolors/issues/63

Hi, thanks for the awesome library!

This is one implementation idea I had to avoid call stack error explained in the above issue. I'm not sure if you have any particular reason you've chosen recursive one (perf? code side?), but just to have something to discuss, I'm starting this PR here.

Please let me know what you think. Thanks!