@alexeyraspopov 2024-05-14
Thanks for raising the issue _and_ making a PR for it! Let me briefly go through the code first. It passes the tests so nothing prevents me from merging it. After the merge I'll need to make some tests to see how it affects existing benchmarks and if we need some more of them for this kind of tasks.