@hi-ogawa 2024-05-14
Very interesting! Right, I was thinking the special casing the first check could have significant impact, but I didn't properly compared with main branch. Thanks for testing with it.

Your code looks good to me and now I updated a branch with your suggestion. Thanks for the review!