@alexeyraspopov 2024-05-14
@hi-ogawa `v1.0.1` is now available in NPM.